Use %llu for all size_t printing. Cast size_t to uint64_t. This will hopefully shutup mingw, which doesn't support c99's %z.

+++ src/mem.c
@@ -137,7 +137,7 @@ static void *Z_Malloc_(size_t size, z_tag_t tag, void *parent) {
 	const size_t s = size + sizeof(z_block_t);
 
 	if (!(z = calloc(s, 1))) {
-		Com_Error(ERR_FATAL, "Failed to allocate %zd bytes\n", s);
+		Com_Error(ERR_FATAL, "Failed to allocate %llu bytes\n", (uint64_t) s);
 	}
 
 	z->magic = Z_MAGIC;
